SLING-8038 - Allow the Repository MOJO specify including features from CLI
initial checkin
